+++ local/base_user.go
@@ -4,8 +4,8 @@ import "frbg/network"
 
 type UserImplement struct {
 	userId uint32
-	gameId uint32
-	gateId uint32
+	gameId uint8
+	gateId uint8
 	*network.Conn
 }
 
@@ -13,10 +13,10 @@ func (u *UserImplement) UserID() uint32 {
 	return u.userId
 }
 
-func (u *UserImplement) GameID() uint32 {
+func (u *UserImplement) GameID() uint8 {
 	return u.gameId
 }
 
-func (u *UserImplement) GateID() uint32 {
+func (u *UserImplement) GateID() uint8 {
 	return u.gateId
 }
